# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Intranet",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/get-branches/{username}": {
            "get": {
                "summary": "Obtene sucursales",
                "operationId": "5fb16c3c5980053ae041a3278b195d73",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "Nombre de usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "Acceso no autorizado"
                    },
                    "500": {
                        "description": "Error en el servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "summary": "Iniciar sesion",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "yordyjc.chura@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login correcto."
                    },
                    "401": {
                        "description": "Contraseña o usuario incorrecto"
                    },
                    "500": {
                        "description": "Error en el servidor"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "summary": "Cerrar sesion",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logged out."
                    },
                    "500": {
                        "description": "Error en el servidor"
                    },
                    "401": {
                        "description": "Tu token no es valido o ya expiro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/user": {
            "get": {
                "summary": "Mostrar usuarios",
                "operationId": "64e612024b8181fcdfd0c0510ac96ce5",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Autenticación con token de Laravel Sanctum (Bearer {token})",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}